SPI: Fixed return values & added SPI end transfers

+++ include/core/spi.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 
 
-uint8_t spi_init(uint8_t mode);
-uint8_t spi_release();
-uint8_t spi_set_speed(uint32_t speed);
+int spi_init(uint8_t mode);
+int spi_release();
+int spi_set_speed(uint32_t speed);
 
-uint8_t spi_transfer(const uint8_t * tx_buffer, uint8_t * rx_buffer, uint32_t len);
+int spi_transfer(const uint8_t * tx_buffer, uint8_t * rx_buffer, uint32_t len);
 
 #endif
